pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sky-agent"
version = "0.1.0"
description = "Sky Agent for multi-cloud infrastructure management"
requires-python = ">=3.11"
dependencies = [
    "strands-agents",
    "strands-agents-tools",
    "mcp",
    "claude-code-sdk",
    "fastapi",
    "uvicorn",
    "pydantic"
]

[project.optional-dependencies]
dev = []

[project.scripts]
sky-agent = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]